remove unused imports in app.js, rename utils.js
